Deprecate and update time-related classes (#34)
With Java 8 a new Time API was introduced. To encourage the use of these we
deprecate our own implementation of `Clock`.

The implementation of `Deadline` was updated to use the new Java Time API. We
chose to use `Instant` internally, as this more accurately expresses the intent.
See
https://stackoverflow.com/questions/32437550/whats-the-difference-between-instant-and-localdatetime.

This change removes `Deadline`-related methods from `Clock`, because they are no
longer compatible. Those who wish to migrate from these methods should use the
constructor for `Deadline`.

Co-authored-by: Ferenc Züllich <ferenc.zuellich@just.social>